event: extract RpkiState and RpkiClient into rpki.rs
Move RPKI protocol logic to daemon/src/rpki.rs, following the same
pattern as bmp.rs and mrt.rs.  The new module depends only on
TableHandle from event.rs and has no dependency on the API layer.

impl From<&RpkiState> for api::RpkiState stays in event.rs because
it converts RPKI state to protobuf types, which belongs in the API
layer rather than the protocol module.
