+++ daemon/src/main.rs
@@ -27,6 +27,7 @@ mod mrt;
 mod peer_tx;
 mod policy;
 mod proto;
+mod rpki;
 mod table_manager;
 
 use clap::{Arg, Command};
